+++ yummy-rs/tests/llm/llm.http
@@ -15,3 +15,13 @@ content-type: application/json
     "input":["Elixir of Eternal Twilight: Grants visions of realms beyond the veil.",
              "Elixir of Eternal Twilight: Grants visions of realms beyond the veil."]
 }
+
+###
+
+POST http://yummy:8080/embeddings HTTP/1.1
+content-type: application/json
+
+{
+    "input":["Elixir of Eternal Twilight: Grants visions of realms beyond the veil.",
+             "Elixir of Eternal Twilight: Grants visions of realms beyond the veil."]
+}
