+++ yummy-rs/yummy-core/Cargo.toml
@@ -10,13 +10,12 @@ license = "Apache-2.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
 byteorder = "1.4.3"
-fasthash = "0.4.0"
 protobuf = "=3.1.0"
 protobuf-json-mapping = "=3.1.0"
 async-trait = "0.1.57"
 tokio = { version = "1.20.0", features = ["full"] }
 thiserror = "1.0"
 regex = "1"
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { version = "0.11", features = ["json", "rustls-tls"], default-features = false}
 url = "2.3"
