+++ yummy-rs/yummy-core/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod common;
 pub mod config;
-pub mod encoding;
