+++ yummy-rs/yummy-delta/Cargo.toml
@@ -30,7 +30,7 @@ log = "0.4.6"
 async-stream = "0.3.3"
 bytes = "1.3.0"
 regex = "1"
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { version = "0.11", features = ["json", "rustls-tls"] }
 serde_json = "1.0.91"
 anyhow = "1.0.69"
 url = "2.3"
