+++ yummy-rs/yummy-features/Cargo.toml
@@ -22,3 +22,4 @@ env_logger = "0.10.1"
 chrono = "0.4.22"
 derive_more = "0.99.17"
 thiserror = "1.0"
+fasthash = "0.4.0"
