+++ yummy-rs/yummy-llm/Cargo.toml
@@ -8,7 +8,7 @@ license = "Apache-2.0"
 
 [dependencies]
 yummy-core = { path = "../yummy-core" }
-actix-web = "4"
+actix-web = { version="4", features=["rustls"]}
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"
 async-trait = "0.1.57"
@@ -31,7 +31,8 @@ tokenizers = { version = "0.13.4", features = ["unstable_wasm"], default-feature
 candle = { git = "https://github.com/huggingface/candle.git", branch = "main", version = "0.3.2", package = "candle-core" }
 candle-nn = { git = "https://github.com/huggingface/candle.git", branch = "main", version = "0.3.2" }
 candle-transformers = { git = "https://github.com/huggingface/candle.git", branch = "main", version = "0.3.2" }
-hf-hub = { version="0.3.0", features=["tokio"]}
+hf-hub = { git = "https://github.com/yummyml/hf-hub.git", branch = "main", version="0.3.0", features=["tokio"] }
+
 
 [features]
 catboost = []
