+++ yummy-rs/yummy/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "yummy"
-version.workspace = true
+version = "0.0.9"
+#version.workspace = true
 edition = "2021"
 repository = "https://github.com/yummyml/yummy"
 description = "The yummy library"
@@ -19,6 +20,7 @@ yummy-delta = { version = "0.0.9",  path = "../yummy-delta", optional = true }
 yummy-ml = { version = "0.0.9",  path = "../yummy-ml", optional = true }
 yummy-llm = { version = "0.0.9",  path = "../yummy-llm", optional = true }
 tokio = { version = "1.20.0", features = ["full"] }
+
 
 [features]
 yummy-delta = ["dep:yummy-delta"]
